+++ lib/model/PrivacyData.dart
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import 'package:rokwire_plugin/service/assets.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
@@ -50,7 +49,7 @@ class PrivacyData{
 
   //Util methods
   String? getLocalizedString(String? text) {
-    return Localization().getStringFromMapping(text, (jsonData != null) ? jsonData!['strings'] : Assets()['privacy.strings']);
+    return Localization().getStringFromMapping(text, (jsonData != null) ? jsonData!['strings'] : null);
   }
 }