+++ lib/ui/groups/GroupsHomePanel.dart
@@ -321,9 +321,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
           InkWell(onTap: _onFilterAttributes, child:
             Padding(padding: EdgeInsets.only(top: 14, bottom: 8), child:
               Row(children: [
-                Text(filtersTitle, style: TextStyle(
-                  fontFamily: Styles().fontFamilies?.bold, fontSize: 16, color: Styles().colors?.fillColorPrimary,
-                ),),
+                Text(filtersTitle, style:  Styles().textStyles?.getTextStyle("widget.title.regular.fat")),
                 Padding(padding: EdgeInsets.symmetric(horizontal: 4), child:
                   Styles().images?.getImage('chevron-right', width: 6, height: 10) ?? Container(),
                 )
@@ -404,7 +402,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
       return Padding(
           padding: EdgeInsets.symmetric(horizontal: 7, vertical: 14),
           child: Text(countLabel,
-              style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 16, color: Styles().colors?.fillColorPrimary)));
+              style: Styles().textStyles?.getTextStyle("widget.title.regular.fat")));
     }
   }
 
@@ -414,7 +412,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
           InkWell(onTap: _onTapCreate, child:
             Padding(padding: EdgeInsets.symmetric(vertical: 10), child:
               Row(crossAxisAlignment: CrossAxisAlignment.center, children: [
-                Text(Localization().getStringEx("panel.groups_home.button.create_group.title", 'Create'), style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 16, color: Styles().colors?.fillColorPrimary)),
+                Text(Localization().getStringEx("panel.groups_home.button.create_group.title", 'Create'), style: Styles().textStyles?.getTextStyle("widget.title.regular.fat")),
                 Padding(padding: EdgeInsets.only(left: 4), child:
                   Styles().images?.getImage('plus-circle', excludeFromSemantics: true)
                 )
@@ -559,7 +557,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
       List<Widget> widgets = [];
       widgets.add(Container(height: 8));
       widgets.add(Padding(padding: EdgeInsets.symmetric(horizontal: 16), child:
-        Text(Localization().getStringEx("panel.groups_home.label.pending", "Pending"), style: TextStyle(fontFamily: Styles().fontFamilies!.bold, fontSize: 20, color: Styles().colors!.fillColorPrimary),)
+        Text(Localization().getStringEx("panel.groups_home.label.pending", "Pending"), style: Styles().textStyles?.getTextStyle("widget.title.large.fat"))
         )
       );
       widgets.add(Container(height: 8,));
@@ -620,7 +618,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
         text = Localization().getStringEx("panel.groups_home.label.all_groups.filtered.empty", "No groups match the selected filter");
       }
       return Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 30), child:
-        Text(text, style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground),),
+        Text(text, style: Styles().textStyles?.getTextStyle("widget.item.regular.thin")),
       );
     }
   }