+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [5.0.4] - 2023-05-22
+### Changed
+- Load different JSON assets from content service [#3278](https://github.com/rokwire/app-flutter-plugin/issues/280).
+- Retire Assets service [#3278](https://github.com/rokwire/app-flutter-plugin/issues/280).
+
+## [5.0.3] - 2023-05-19
+### Changed
+- Use "content_id" as guide article identifier [#3287](https://github.com/rokwire/illinois-app/issues/3287).
+- Use Styles().textStyles everywhere [#2932](https://github.com/rokwire/illinois-app/issues/2932.)
+
+## [5.0.2] - 2023-05-17
 ### Added
 - New video tutorial "Creating Polls" [#3285](https://github.com/rokwire/illinois-app/issues/3285).